+++ CHANGELOG.md
@@ -24,6 +24,7 @@ en las distintas versiones de la misma.
 * feat: Permitir añadir información adicional al evaluar el desempeño de un estudiante de dual
 * feat: Incluir más tipos de preguntas en las encuestas (números del 0-5 y 0-10)
 * feat: Mostrar el número de convenio de FP dual en el listado del proyecto
+* feat: Generación de un informe de contactos para responsables de seguimiento de FP dual
 
 5.0.0 (2022-04-20)
 ------------------